[7905] Amended performance profiles to be on its own
