+++ app/views/guidance/performance_profiles.html.erb
@@ -32,7 +32,7 @@
     <ul class="govuk-list govuk-list--bullet">
       <li>
         downloading a
-        <a href=<%= performance_profiles_reports_path %> class="govuk-link">report with the data you need to check</a>
+        <a href=<%= reports_performance_profiles_path %> class="govuk-link">report with the data you need to check</a>
       </li>
       <li>
         exporting a
